Marking TrueTypeFont based font implementations as deprecated, fitting the Slick implementation Changing Unicode implementation name to reflect the use of Hiero settings
